backend/requirements.txt
fastapi===0.111.0
uvicorn===0.29.0
python-dotenv===1.0.1
PyGithub===2.3.0
langchain_google_genai===1.0.4
langchain===0.1.20